# Restrict fastpath isel indexes to the case of all PandasIndex