<!-- Feel free to remove check-list items aren't relevant to your change -->

- [x] Closes #10063
- [x] Tests added
- [x] User visible changes (including notable bug fixes) are documented in `whats-new.rst`

Implement https://github.com/pydata/xarray/issues/10063#issuecomment-2672005278. Since we use `type(idx) is ...` instead of `isinstance(idx, ...)` this should be even faster.

The only performance degradation would be in the case of single coordinate custom indexes implemented in 3rd-party libraries, which could go through the fastpath as well but is harder to detect, but I think that's ok for now.

@hmaarrfk @dcherian