@benbovy 2025-02-24
Looking at your benchmark, I think we could (should?) cache `Dataset.xindexes`, which creates one temporary dict and two other dicts for storing the indexes and variables in the `Indexes` object and which also does redundant checks. We would just need to ensure that in-place updating of the Dataset indexed coordinates will invalidate the cache.

We could also cache `Indexes.group_by_index()` (safe here since `Indexes` is immutable). In repeated calls most time is spent on creating the list and dict objects, since `Indexes._id_coord_names` is already cached.

I think this will provide benchmark results at least on par with the fastest ones above while this will work with all kinds of indexes (both single and multi-coordinate) and without duplicating `_apply_indexes` logic.